Move config.json to root
