+++ configTemplate.json
